+++ examples/pysuccinct/pyfile.py
@@ -1,4 +1,4 @@
-import pysemistructured
+import pyfile
 import sys
 import getopt
 
@@ -30,7 +30,7 @@ else:
     if (option[0] == "load"):
         # Load file from memory
         print("loading ", inputpath, " from file")
-        q = pysemistructured.PySemistructured(inputpath)
+        q = pyfile.PyFile(inputpath)
     elif (option[0] == "compress"):
         # Compress the file
         print("Please enter the sampling rates")
@@ -57,41 +57,42 @@ else:
             else:
                 printf("Invalid Option")
                 sys.exit(2)
-        q = pysemistructured.PySemistructured(inputpath, sa_sampling_rate, 
+        q = pyfile.PyFile(inputpath, sa_sampling_rate, 
             isa_sampling_rate, npa_sampling_rate, 
             sampling_scheme, npa_encoding_scheme)
     else:
         print("Usage: [load/compress] [file]\n")
         sys.exit(2)
 
-print("Command must be one of:\n\t\tsearch [attr_key] [attr_val]\n\t\tcount [attr_key] [attr_val]\n\t\tget [key] [attr_key]")
-
 # Parse through line by line
 while (True):
     line = input("succinct> ")
-    line = line.split(" ")
+    line = line.split(" ", 1)
     if (line[0] == "search"):
-        if (len(line) != 3):
+        if (len(line) != 2):
             print("Could not parse command: ")
+            continue
         else:
-            print(q.PySearch(line[1].strip(), line[2].strip()))
+            print(q.Search(line[1].strip()))
     elif(line[0] == "count"):
-        if (len(line) != 3):
+        if (len(line) != 2):
             print("Could not parse command: ")
+            continue
         else:
-            print(q.Count(line[1].strip(), line[2].strip()))
-    elif(line[0] == "get"):
+            print(q.Count(line[1].strip()))
+    elif(line[0] == "extract"):
         if (len(line) == 1):
             print("Could not parse command: ")
             continue
-        key = line[1].strip()
-        attr_key = line[2].strip()
-        if (len(line) != 3 or RepresentsInt(key) == False):
+        line = line[1].split(" ", 1)
+        offset = line[0].strip()
+        length = line[1].strip()
+        if (len(line) != 2 or RepresentsInt(offset) == False or RepresentsInt(length) == False):
             print("Could not parse command: ")
         else:
-            print(q.Get((int(key), attr_key)))
+            print(q.Extract(int(line[0].strip()), int(line[1].strip())))
     elif(line[0] == "exit"):
         break
     else:
         print("Unsupported command")
-        print("Command must be one of:\n\t\tsearch [attr_key] [attr_val]\n\t\tcount [attr_key] [attr_val]\n\t\tget [key] [attr_key]")
+        print("Command must be one of: search [query], count [query], extract [offset] [length]")
